move getPermissionNameAbbreviation, getTemplatePermissionAbbreviation, remapObjectKeysWithPrefix to helpersUtils
