+++ src/components/permissions/rowLevelPolicyHelpers.ts
@@ -1,31 +1,13 @@
 import databaseConnect from '../databaseConnect'
 import { UserInfo, PermissionRow, RuleTypes } from './types'
+import {
+  getPermissionNameAbbreviation,
+  getTemplatePermissionAbbreviation,
+  remapObjectKeysWithPrefix,
+} from './helpersUtils'
 
 const SQLBuilder = require('json-sql-builder2')
 const jsonToSql = new SQLBuilder('PostgreSQL')
-
-// in { permissionPolicyId: 1, permissionNameId:2 }
-// out "pp1pn2"
-const getPermissionNameAbbreviation = ({ permissionPolicyId, permissionNameId }: PermissionRow) =>
-  `pp${permissionPolicyId}pn${permissionNameId}`
-
-// in { permissionPolicyId: 1, permissionNameId:2, templatePermissionId: 3 }
-// out "pp1pn2tp3"
-const getTemplatePermissionAbbreviation = (permissionRow: PermissionRow) =>
-  `${getPermissionNameAbbreviation(permissionRow)}tp${permissionRow.templatePermissionId}`
-
-// Returns new version of object with each key prefixed with prefix (and value converter to string)
-// in "prefixMe", { one: 1, two: 2}
-// out { prefixMe_one: "1", prefixMe_two: "2" }
-const remapObjectKeysWithPrefix = (prefix: string, object: Object) => {
-  const remappedObject: { [index: string]: string } = {}
-
-  Object.entries(object).forEach(([key, value]) => {
-    remappedObject[`${prefix}_${key}`] = String(value)
-  })
-
-  return remappedObject
-}
 
 /* Compiles JWT from userInfo and PerissionRows
   in { userId: 1, ... }, [
